moved Icons.js into CardLayout
