+++ src/ReactBusinessCard/CardLayout/index.js
@@ -1,6 +1,6 @@
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
-import { GithubIcon, LinkedinIcon, ResumeIcon } from '../Icons'
+import { GithubIcon, LinkedinIcon, ResumeIcon } from './Icons'
 import './styles.css'
 
 const downloadResume = (uri, first, last) => {
